requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.12.0=pypi_0
argon2-cffi=20.1.0=pypi_0
ase=3.21.1=pypi_0
async-generator=1.10=pypi_0
attrs=20.3.0=pypi_0
backcall=0.2.0=py_0
blas=1.0=mkl
bleach=3.2.3=pypi_0
boto3=1.17.29=pypi_0
botocore=1.20.29=pypi_0
ca-certificates=2020.12.5=ha878542_0
cached-property=1.5.2=pypi_0
cachetools=4.2.1=pypi_0
certifi=2020.12.5=py36h5fab9bb_1
cffi=1.14.4=pypi_0
chardet=4.0.0=pypi_0
click=7.1.2=pypi_0
configargparse=1.4=pypi_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.10.0=py36_0
dataclasses=0.7=py36_0
dbus=1.13.18=hb2f20db_0
decorator=4.4.2=py_0
deepdish=0.3.6=pypi_0
defusedxml=0.6.0=pypi_0
entrypoints=0.3=pypi_0
expat=2.2.10=he6710b0_2
filelock=3.0.12=pypi_0
fontconfig=2.13.0=h9420a91_0
freetype=2.10.4=h5ab3b9f_0
gensim=3.8.3=pypi_0
git-lfs=2.13.3=ha770c72_0
glib=2.66.1=h92f7085_0
google-auth=1.28.0=pypi_0
google-auth-oauthlib=0.4.4=pypi_0
googledrivedownloader=0.4=pypi_0
grpcio=1.36.1=pypi_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
h5py=3.1.0=pypi_0
icu=58.2=he6710b0_3
idna=2.10=pypi_0
importlib-metadata=3.4.0=pypi_0
intel-openmp=2020.2=254
ipykernel=5.3.4=py36h5ca1d4c_0
ipython=7.16.1=py36h5ca1d4c_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.3=pypi_0
isodate=0.6.0=pypi_0
jedi=0.18.0=py36h06a4308_1
jinja2=2.11.3=pypi_0
jmespath=0.10.0=pypi_0
joblib=1.0.0=pypi_0
jpeg=9b=h024ee3a_2
jsonschema=3.2.0=pypi_0
jupyter=1.0.0=pypi_0
jupyter-console=6.2.0=pypi_0
jupyter_client=6.1.7=py_0
jupyter_core=4.7.0=py36h06a4308_0
jupyterlab-pygments=0.1.2=pypi_0
jupyterlab-widgets=1.0.0=pypi_0
jupyterthemes=0.20.0=py_1
kiwisolver=1.3.0=py36h2531618_0
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
lesscpy=0.13.0=py_1
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuuid=1.0.3=h1bed415_2
libuv=1.40.0=h7b6447c_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=hb55368b_3
llvmlite=0.36.0=pypi_0
lz4-c=1.9.3=h2531618_0
markdown=3.3.4=pypi_0
markupsafe=1.1.1=pypi_0
matplotlib=3.3.2=h06a4308_0
matplotlib-base=3.3.2=py36h817c723_0
mistune=0.8.4=pypi_0
mkl=2020.2=256
mkl-service=2.3.0=py36he8ac12f_0
mkl_fft=1.2.0=py36h23d657b_0
mkl_random=1.1.1=py36h0573a6f_0
nbclient=0.5.1=pypi_0
nbconvert=6.0.7=pypi_0
nbformat=5.1.2=pypi_0
ncurses=6.2=he6710b0_1
nest-asyncio=1.5.1=pypi_0
networkx=2.5=pypi_0
ninja=1.10.2=py36hff7bd54_0
nlp-python=1.1.0=pypi_0
nltk=3.5=pypi_0
notebook=6.2.0=pypi_0
numba=0.53.1=pypi_0
numexpr=2.7.3=pypi_0
numpy=1.19.2=py36h54aff64_0
numpy-base=1.19.2=py36hfa32c7d_0
oauthlib=3.1.0=pypi_0
olefile=0.46=py36_0
openssl=1.1.1k=h27cfd23_0
packaging=20.9=pypi_0
pandas=1.1.3=py36he6710b0_0
pandocfilters=1.4.3=pypi_0
parso=0.8.1=pyhd3eb1b0_0
pcre=8.44=he6710b0_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.1.0=py36he98fc37_0
pip=20.3.3=py36h06a4308_0
ply=3.11=py_1
prometheus-client=0.9.0=pypi_0
prompt-toolkit=3.0.8=py_0
protobuf=3.15.7=pypi_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.20=pypi_0
pygments=2.7.4=pyhd3eb1b0_0
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py36h05f1152_2
pyrsistent=0.17.3=pypi_0
python=3.6.12=hcff3b4d_2
python-dateutil=2.8.1=py_0
python-louvain=0.15=pypi_0
python_abi=3.6=1_cp36m
pytorch=1.7.1=py3.6_cuda10.2.89_cudnn7.6.5_0
pytz=2021.1=pyhd3eb1b0_0
pyzmq=20.0.0=py36h2531618_1
qt=5.9.7=h5867ecd_1
qtconsole=5.0.2=pypi_0
qtpy=1.9.0=pypi_0
rdflib=5.0.0=pypi_0
readline=8.0=h7b6447c_0
regex=2020.11.13=pypi_0
requests=2.25.1=pypi_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.7.2=pypi_0
s3transfer=0.3.4=pypi_0
sacremoses=0.0.43=pypi_0
scikit-learn=0.24.1=pypi_0
scipy=1.5.4=pypi_0
send2trash=1.5.0=pypi_0
sentencepiece=0.1.95=pypi_0
setuptools=51.3.3=py36h06a4308_4
sip=4.19.8=py36hf484d3e_0
six=1.15.0=py36h06a4308_0
smart-open=4.1.2=pypi_0
sqlite=3.33.0=h62c20be_0
tables=3.6.1=pypi_0
tensorboard=2.4.1=pypi_0
tensorboard-plugin-wit=1.8.0=pypi_0
terminado=0.9.2=pypi_0
testpath=0.4.4=pypi_0
threadpoolctl=2.1.0=pypi_0
tk=8.6.10=hbc83047_0
tokenizers=0.10.1=pypi_0
torch-cluster=1.5.9=pypi_0
torch-geometric=1.6.3=pypi_0
torch-scatter=2.0.6=pypi_0
torch-sparse=0.6.9=pypi_0
torch-spline-conv=1.2.1=pypi_0
torchaudio=0.7.2=py36
torchvision=0.8.2=py36_cu102
tornado=6.1=py36h27cfd23_0
tqdm=4.56.0=pypi_0
traitlets=4.3.3=py36_0
transformers=4.4.1=pypi_0
typing_extensions=3.7.4.3=py_0
urllib3=1.26.4=pypi_0
wcwidth=0.2.5=py_0
webencodings=0.5.1=pypi_0
werkzeug=1.0.1=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
widgetsnbextension=3.5.1=pypi_0
xz=5.2.5=h7b6447c_0
zeromq=4.3.3=he6710b0_3
zipp=3.4.0=pypi_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
